backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg[binary]==3.1.19
pydantic==2.7.3
pydantic-settings==2.3.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
alembic==1.13.2
loguru==0.7.2
websockets==12.0
Pillow==10.4.0
aiofiles==24.1.0
reportlab==4.0.9


